requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
requests>=2.26.0,<3.0.0
python-dotenv>=0.19.0,<0.20.0
redis>=4.0.0,<5.0.0
prometheus-client>=0.11.0,<0.12.0
starlette-exporter>=0.11.0,<0.12.0
aiohttp>=3.8.0,<4.0.0